requirements.txt
fastapi
uvicorn
supabase-py
librosa
soundfile
requests
